Finished core functions for the first script in the dropdown menu
Created a class to pass to the different methods being used instead of the 6 different variables I was using at the time.

Changed the name of scriptSelector.py to AddTextToStartOfCell.py and added it into a subfolder called scriptFiles. scriptFiles will hold the .py's for the other scripts to be added.

changed UploadTableHelper.py to end in .pyw to open without console.

Added comments to the class to explain what each variable is.